Refactor: 게시글 작성 페이지 코드 분리 및 디자인 변경
