+++ src/Components/index.tsx
@@ -5,7 +5,6 @@ export * from './FeedCard/CardList';
 export * from './FeedCard/Card';
 export * from './Feed/FeedDetail';
 export * from './Feed/Comment/CommentContainer';
-export * from './Feed/WriteFeed';
 export * from './Channel/MyFeeds';
 export * from './Channel/MyComments';
 export * from './CheckPassword';
